Did some cleaning up and added some event modifiers
This version introduces a bunch of other methods of uploading and getting images, although they are not functional yet.
Reason being i have to clean up my old functions, which means rewriting them in a more modular and easy to read and modify fassion. for now the main start function is completely redone. and two new functions handle orientation and screen capture. i will work more on this in the upcoming days.